refactor(chatBox): split chatBox into a component & container
